CarDealership: add loading skeleton UI
